requirements.txt
aiohttp==3.7.4
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clr-loader==0.2.6
contourpy==1.2.1
cycler==0.12.1
Django==4.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
fonttools==4.52.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==2.10
iso8601==2.1.0
kiwisolver==1.4.5
kombu==5.3.7
matplotlib==3.9.0
metaapi_cloud_copyfactory_sdk==10.0.0
metaapi_cloud_metastats_sdk==5.0.0
metaapi_cloud_sdk==27.0.5
mt4pycon==1.0.2
multidict==4.7.6
numpy==1.26.4
ohtaylor==1.1.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
prompt_toolkit==3.0.45
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==1.7.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==3.14.2
python-socketio==4.6.0
pythonnet==3.0.3
pytz==2024.1
rapidfuzz==3.9.0
redis==5.0.4
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==1.25.11
vine==5.1.0
wcwidth==0.2.13
websockets==11.0.3
yarl==1.9.4
